Fixed problem siamese example
The custom loss required labels to be in float32 type. I have changed the "generator" (create_pairs) to output the expected type

+++ examples/mnist_siamese.py
@@ -64,7 +64,7 @@ def create_pairs(x, digit_indices):
             z1, z2 = digit_indices[d][i], digit_indices[dn][i]
             pairs += [[x[z1], x[z2]]]
             labels += [1, 0]
-    return np.array(pairs), np.array(labels)
+    return np.array(pairs), np.array(labels).astype('float32')
 
 
 def create_base_network(input_shape):
